backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
supabase-py==2.3.0
